Split expectations in different files
